+++ src/components/dialog/dialog.component.tsx
@@ -5,6 +5,7 @@ import React, {
   useCallback,
   useImperativeHandle,
   forwardRef,
+  useState,
 } from "react";
 
 import createGuid from "../../__internal__/utils/helpers/guid";
@@ -140,6 +141,10 @@ export const Dialog = forwardRef<DialogHandle, DialogProps>(
     const containerRef = useRef<HTMLDivElement>(null);
     const innerContentRef = useRef(null);
     const titleRef = useRef(null);
+    const [breakpointOffset, setBreakpointOffset] = useState<
+      number | undefined
+    >(undefined);
+    const isDialogMaximised = size === "maximise";
     const listenersAdded = useRef(false);
     const { current: titleId } = useRef(createGuid());
     const { current: subtitleId } = useRef(createGuid());
@@ -181,9 +186,16 @@ export const Dialog = forwardRef<DialogHandle, DialogProps>(
         midPointX = 0;
       }
 
+      if (isDialogMaximised) {
+        const breakPoint = window.innerWidth > 960 ? 32 : 16;
+        midPointX = breakPoint;
+        midPointY = breakPoint;
+        setBreakpointOffset(breakPoint);
+      }
+
       containerRef.current.style.top = `${midPointY}px`;
       containerRef.current.style.left = `${midPointX}px`;
-    }, []);
+    }, [size]);
 
     useResizeObserver(innerContentRef, centerDialog, !open);
 
@@ -308,7 +320,11 @@ export const Dialog = forwardRef<DialogHandle, DialogProps>(
             data-role={dataRole}
             aria-modal={isTopModal ? true : undefined}
             ref={containerRef}
-            topMargin={TOP_MARGIN}
+            topMargin={
+              isDialogMaximised && breakpointOffset
+                ? breakpointOffset * 2
+                : TOP_MARGIN
+            }
             {...dialogProps}
             role={role}
             tabIndex={-1}