+++ src/components/loader-star/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./loader-star.component";
+export type { LoaderStarProps } from "./loader-star.component";