+++ src/components/tile/tile.style.ts
@@ -3,10 +3,17 @@ import { space, SpaceProps } from "styled-system";
 import baseTheme from "../../style/themes/base";
 import computeSizing from "../../style/utils/element-sizing";
 import { TileProps } from "./tile.component";
+import StyledTileContent from "./tile-content/tile-content.style";
 
 type StyledTileProps = Pick<
   TileProps,
-  "borderWidth" | "borderVariant" | "variant" | "width" | "roundness" | "height"
+  | "borderWidth"
+  | "borderVariant"
+  | "variant"
+  | "width"
+  | "roundness"
+  | "height"
+  | "highlightVariant"
 > & { isHorizontal?: boolean } & SpaceProps;
 
 const getBorderColor = (
@@ -47,6 +54,24 @@ const getBorderRadius = (roundness: TileProps["roundness"]) => {
   }
 };
 
+const getHeighlightVariant = (variant: TileProps["highlightVariant"]) => {
+  switch (variant) {
+    case "success":
+      return "var(--colorsSemanticPositive500)";
+    case "neutral":
+      return "var(--colorsSemanticNeutral500)";
+    case "error":
+      return "var(--colorsSemanticNegative500)";
+    case "warning":
+      return "var(--colorsSemanticCaution500)";
+    case "info":
+      return "var(--colorsSemanticInfo500)";
+    default:
+      // gradient is default
+      return "linear-gradient(0deg, rgb(143, 73, 254) 5%, rgb(0, 146, 219) 50%, rgb(19, 160, 56) 95%)";
+  }
+};
+
 const StyledTile = styled.div<StyledTileProps>`
   ${({
     borderVariant,
@@ -65,12 +90,12 @@ const StyledTile = styled.div<StyledTileProps>`
     border-radius: ${getBorderRadius(roundness)};
     --tileBorderRadius: ${getBorderRadius(roundness)};
 
-    > *:first-child {
+    > *:first-child:not(${StyledTileContent}) {
       border-top-left-radius: calc(${getBorderRadius(roundness)} - 1px);
       border-top-right-radius: calc(${getBorderRadius(roundness)} - 1px);
     }
 
-    > *:last-child {
+    > *:last-child:not(${StyledTileContent}) {
       border-bottom-left-radius: calc(${getBorderRadius(roundness)} - 1px);
       border-bottom-right-radius: calc(${getBorderRadius(roundness)} - 1px);
     }
@@ -108,5 +133,22 @@ const StyledTile = styled.div<StyledTileProps>`
 StyledTile.defaultProps = {
   theme: baseTheme,
 };
+
+export const StyledHighlight = styled.div<{
+  variant: Required<TileProps["highlightVariant"]>;
+  roundness: TileProps["roundness"];
+}>`
+  height: 100%;
+  width: 100%;
+  position: relative;
+  background: ${({ variant }) => getHeighlightVariant(variant)};
+  border-radius: ${({ roundness }) => getBorderRadius(roundness)};
+
+  ${StyledTile} {
+    border-left: 0;
+    left: 4px;
+    width: calc(100% - 4px);
+  }
+`;
 
 export default StyledTile;